Support for network strategy param (#435)
* [multichain] Use blockfinder to get block nums

* Add potion strategy

* Add rarity score

* Add more pages

* Support for network strategy param
